Add lots of files from other branch in new project structure

+++ setup.sh
@@ -1,17 +1,23 @@
 #!/bin/bash
 
-# Activate virtual environment or create one if it 
-#   does not yet exist
-if [ -d "venv" ]; then
-    source venv/Scripts/activate
-else
-    echo "Creating python virtual environment called 'venv'"
-    python -m venv venv
-    source venv/Scripts/activate 
+VENV_DIR="venv"
+
+# Check if the virtual environment directory exists
+if [ ! -d "$VENV_DIR" ]; then
+    echo "Creating virtual environment..."
+    python -m venv $VENV_DIR
 fi
 
-echo "Installing project dependencies..."
-pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
-pip install -r requirements.txt
-echo "Dependencies installed."
+# Activate the virtual environment
+source $VENV_DIR/bin/activate
 
+# Install dependencies from requirements.txt
+if [ -f "requirements.txt" ]; then
+    echo "Installing dependencies..."
+    pip install -r requirements.txt
+else
+    echo "No requirements.txt file found."
+fi
+
+# Deactivate the virtual environment
+deactivate
